+++ xc/lib/Xt/Callback.c
@@ -85,7 +85,7 @@ static InternalCallbackList* FetchInternalList(widget, name)
     offsets = (CallbackTable) 
 	widget->core.widget_class->core_class.callback_private;
 
-    for (n = (int) *(offsets++); --n >= 0; offsets++)
+    for (n = (int)(long) *(offsets++); --n >= 0; offsets++)
 	if (quark == (*offsets)->xrm_name) {
 	    retval = (InternalCallbackList *) 
 		((char *) widget - (*offsets)->xrm_offset - 1);
